refactor(runtime): rename tool_result_content to tool_result_replay
The module's job is replay normalization, not content in general; the new name
says so. Pure rename plus the four import lines that referenced it, with the
mirrored test module renamed to match. File contents are byte-identical.

The public tool_result_content_blocks() function keeps its name.
